Create sdl module and event module.
