+++ src/CMakeLists.txt
@@ -1,9 +1,11 @@
+add_subdirectory(events)
 add_subdirectory(traceLog)
 add_subdirectory(types)
 add_subdirectory(ui)
-add_subdirectory(app)
-add_subdirectory(core)
-add_subdirectory(camera)
-add_subdirectory(image)
+add_subdirectory(sdl)
 add_subdirectory(settings)
+add_subdirectory(image)
+add_subdirectory(camera)
+add_subdirectory(core)
+add_subdirectory(app)
 add_subdirectory(tests)
