+++ src/core/Core/Application.cpp
@@ -1,8 +1,8 @@
 #include "Application.hpp"
 #include "Camera/ICamera.h"
-#include "DpiHandler.h"
-#include "FrameContext.h"
-#include "SdlCameraTexture.h"
+#include "Sdl/DpiHandler.h"
+#include "Sdl/FrameContext.h"
+#include "Sdl/SdlCameraTexture.h"
 #include "TraceLog/Log.hpp"
 #include "TraceLog/Tracy.hpp"
 #include "Ui/IAppUi.h"
