+++ src/events/Events/Bus.h
@@ -1,0 +1,21 @@
+#pragma once
+#include "BusImpl.h"
+#include "Events.h"
+
+#include <variant>
+
+namespace App
+{
+    /** Type alias for the variant of all events used in the application. */
+    using Events = std::variant<
+        EventQuit,
+        EventClose,
+        EventMinimized,
+        EventShown,
+        EventRestored,
+        EventDisplayChanged
+        >;
+
+    /** Type alias for the event bus used in the application, based on the defined events. */
+    using EventBus = EventBusImpl<Events>;
+}
