+++ src/sdl/Sdl/SdlCameraTexture.h
