+++ src/sdl/Sdl/src/SdlCameraTexture.cpp
